Slight size reduction by use of smallints/consts:
- 288016           1924    7176  297116   4889c libuClibc-0.9.29.so
+ 287956           1924    7168  297048   48858 libuClibc-0.9.29.so

+++ libc/inet/getnetbyad.c
@@ -23,7 +23,7 @@ libc_hidden_proto(setnetent)
 libc_hidden_proto(getnetent)
 libc_hidden_proto(endnetent)
 
-extern int _net_stayopen attribute_hidden;
+extern smallint _net_stayopen attribute_hidden;
 
 struct netent *getnetbyaddr (uint32_t net, int type)
 {
