+++ libc/inet/getnetbynm.c
@@ -25,7 +25,7 @@ libc_hidden_proto(setnetent)
 libc_hidden_proto(getnetent)
 libc_hidden_proto(endnetent)
 
-extern int _net_stayopen attribute_hidden;
+extern smallint _net_stayopen attribute_hidden;
 
 struct netent *
 getnetbyname(const char *name)
